new gradle, junit5, reports, and github actions

+++ src/main/java/project_euler/math/Combinatorics.java
@@ -14,7 +14,7 @@ public class Combinatorics {
         this.factorial = factorial;
     }
 
-    public long binominalCoefficient(int n, int k) {
+    public long binominalCoefficient(long n, long k) {
         return factorial(n).divide(factorial(k).multiply(factorial(n - k))).longValue();
     }
